refactor: migrate loader engine into domain package
